Move agg funcs to new file
